requirements.txt
gradio>=4.0.0
Pillow>=10.0.0
requests>=2.31.0
numpy>=1.24.0
opencv-python>=4.8.0
rembg>=2.0.50
onnxruntime>=1.15.0
torch>=2.0.0
torchvision>=0.15.0
